feat(api) add `?check_hash=1` query argument to `/config`
This introduces a new query argument so that the hash of a given
declarative configuration is checked against the hash of the current
configuration before loading it.

```
$ http :8001/config?check_hash=1 config=@config.yml   # load a config
201 Created
{
   ...
}

$ http :8001/config?check_hash=1 config=@config.yml   # reload the same config
304 Not Modified
```

Hash checking only happens when the value of `check_hash` is `1`.
(This design was chosen to allow for future extensibility, if we
add new hashing behaviors, those can be set to different numbers.)